refactor: 홈페이지에 바뀐 와플카드 컴포넌트로 변경
